// src/types/schedule.ts
// Generated by https://quicktype.io

export interface Schedule {
  schedule: ScheduleClass;
}

export interface ScheduleClass {
  version: string;
  base_url: string;
  conference: Conference;
}

export interface Conference {
  acronym: string;
  title: string;
  start: string;
  end: string;
  daysCount: number;
  timeslot_duration: string;
  rooms: RoomElement[];
  days: Day[];
}

export interface Day {
  index: number;
  date: string;
  day_start: string;
  day_end: string;
  rooms: Rooms;
}

export interface Rooms {
  'Magenta Room': Room[];
  'Khaki Room': Room[];
}

export interface Room {
  id: number;
  guid: string;
  logo: string;
  date: string;
  start: string;
  duration: Duration;
  room: Name;
  slug: string;
  url: string;
  title: string;
  subtitle: string;
  track: Track;
  type: Type;
  language: Language;
  abstract: string;
  description: string;
  recording_license: string;
  do_not_record: boolean;
  persons: Person[];
  links: any[];
  attachments: any[];
  answers: any[];
}

export enum Duration {
  The0030 = '00:30',
  The0130 = '01:30',
  The0330 = '03:30',
}

export enum Language {
  En = 'en',
}

export interface Person {
  id: number;
  code: string;
  public_name: string;
  biography: string;
  answers: any[];
}

export enum Name {
  KhakiRoom = 'Khaki Room',
  MagentaRoom = 'Magenta Room',
}

export enum Track {
  FrontLine = 'Front-line',
  Realigned = 'Realigned',
}

export enum Type {
  Talk = 'Talk',
  Workshop = 'Workshop',
}

export interface RoomElement {
  name: Name;
  guid: null;
  description: null;
  capacity: null;
}